use of long long int

+++ demehin.maxim/S2/expr_convert.hpp
@@ -10,8 +10,8 @@ namespace demehin
   using ExprsStack = demehin::Stack< ExprQueue >;
   void convertStack(ExprsStack& infExprs, ExprsStack& postExprs);
   ExprQueue convertQueue(ExprQueue& infExpr);
-  int calculateExpr(const ExprQueue& expr);
-  void getExprsValues(ExprsStack postStack, Stack< int >& values);
+  long long int calculateExpr(const ExprQueue& expr);
+  void getExprsValues(ExprsStack postStack, Stack< long long int >& values);
 }
 
 #endif
